requirements.txt
python-docx
markdown
beautifulsoup4